initial db read call implemented, configuration tbd

+++ source/Domain.RentalInventory/Domain.RentalInventory.WebApi/Controllers/RentalPropertyController.cs
@@ -30,7 +30,7 @@ namespace Domain.RentalInventory.WebApi.Controllers
 
         // GET api/<controller>/5
         [HttpGet("{id}")]
-        public async Task<RentalPropertyViewModel> Get(int id)
+        public async Task<RentalPropertyViewModel> Get(long id)
         {
             var request = new GetRentalPropertyByIdRequest(id);
